refactor: Form 클래스 생성 및 initForm을 RacingGame 메소드로 변경
