+++ src/index.js
@@ -1,8 +1,7 @@
 import RacingGame from './RacingGame.js';
-import initHTML from './utils/initializers/initHTML.js';
-import initForms from './utils/initializers/initForms.js';
+import initHTML from './utils/initHTML.js';
 
 initHTML();
 
 const racingGame = new RacingGame();
-initForms(racingGame);
+racingGame.initForm();
